feat: OpenFeign 기반 예매 결제 시스템 연동

+++ src/main/java/com/ceos22/cgvclone/domain/theater/DTO/ShowtimeListResponseDTO.java
@@ -3,6 +3,7 @@ package com.ceos22.cgvclone.domain.theater.DTO;
 import com.ceos22.cgvclone.domain.theater.entity.Showtime;
 import com.ceos22.cgvclone.domain.theater.enums.ScreenType;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -11,7 +12,7 @@ public record ShowtimeListResponseDTO(
         String movieTitle,
         ScreenType screenType,
         String screenName,
-        int price,
+        BigDecimal price,
         LocalDateTime startTime,
         LocalDateTime endTime
 ) {}
